Rename files to match rest of repo
