+++ src/app/components/sites/pages/details/details.component.ts
