+++ src/app/components/sites/pages/harvest/harvest.component.spec.ts
@@ -3,7 +3,7 @@ import { ComponentFixture, TestBed } from "@angular/core/testing";
 import { RouterTestingModule } from "@angular/router/testing";
 import { MockBawApiModule } from "@baw-api/baw-apiMock.module";
 import { SharedModule } from "@shared/shared.module";
-import { SiteHarvestComponent } from "./site.component";
+import { SiteHarvestComponent } from "./harvest.component";
 
 describe("SiteHarvestComponent", () => {
   let component: SiteHarvestComponent;
