+++ src/app/components/sites/sites.module.ts
@@ -5,11 +5,11 @@ import { getRouteConfigForPage } from "@helpers/page/pageRouting";
 import { MapModule } from "@shared/map/map.module";
 import { SharedModule } from "@shared/shared.module";
 import { SiteComponent } from "./components/site/site.component";
-import { SiteDeleteComponent } from "./pages/delete/site.component";
-import { SiteDetailsComponent } from "./pages/details/site.component";
-import { SiteEditComponent } from "./pages/edit/site.component";
-import { SiteHarvestComponent } from "./pages/harvest/site.component";
-import { SiteNewComponent } from "./pages/new/site.component";
+import { SiteDeleteComponent } from "./pages/delete/delete.component";
+import { SiteDetailsComponent } from "./pages/details/details.component";
+import { SiteEditComponent } from "./pages/edit/edit.component";
+import { SiteHarvestComponent } from "./pages/harvest/harvest.component";
+import { SiteNewComponent } from "./pages/new/new.component";
 import { WizardComponent } from "./pages/wizard/wizard.component";
 import { pointsRoute } from "./points.menus";
 import { sitesRoute } from "./sites.menus";
